requirements.txt
streamlit-tree-select>=0.0.5
pandas
matplotlib
seaborn
scikit-learn
sentence-transformers
openai>=1.0.0
python-dotenv
umap-learn
plotly
transformers>=4.40.0
torch>=2.1.0
tiktoken
keybert
wordcloud
textblob
psutil
fpdf
streamlit-antd-components>=0.3.0
